Change Syllable To TPhone and have bugs fixed.
It was a mistake to describe the transition between CVS symbols as
"Syllable" and the list as "SyllableList". I guess "TPhone" is a better
name so I changed all "Syllalbe" to "TPhone". TPhone represents
Transiting Phone.
It might be difficult if I change it later because the growing size of
codes -> more efforts on coding and debuging...
I fixed a bug in Period Prediction, a function that helps calculating
the MixRatio in the beginning and ending of a frame to keep the
beginning and ending fixing with the adjacent frames. The bug is the
possibility that endratio is smaller than the startratio.